Cohort evaluation

Run stochastic individual model across trial cohort individuals in one or more clusters (alternate version incorporating more details of testing and reactive treatment during trial, in progress)

cohort(
  mainpop_data = list(),
  cluster_data = list(),
  n_patients = 1,
  prop_T_c = 0.9,
  age_start = 0.5,
  age_end = 10,
  flag_output = 0,
  test_time_values = c(),
  test_type = "clin",
  flag_pre_clearing = 1,
  censor_period = 0,
  flag_reactive_treatment = 1
)

Arguments

mainpop_data
List output by mainpop() containing main population data

cluster_data
List output by clusters_create() containing cluster data

n_patients
Number of trial cohort patients per cluster

prop_T_c
Treatment probability in cohort

age_start
Minimum age of trial cohort patients

age_end
Maximum age of trial cohort patients

flag_output
Integer indicating whether to show progress of simulation

test_time_values
Time points at which tests are administered

test_type
Type of test administered ("clin" = clinical, "RDT" = rapid diagnostic test)

flag_pre_clearing
Integer indicating whether patients given pre-trial prophylaxis (if 1, place all patients into prophylaxis category at start)

censor_period
Time period after a positive test during which a patient is not counted towards incidence

flag_reactive_treatment
Integer indicating whether patients are automatically given prophylaxis after a positive test (shifting them into treatment category if a clinical case, prophylaxis category otherwise)

Details

Takes a list of parameters, returns a list of raw data (data also saved to files as backup).