Power calculation

Function which takes in trial parameters for main population modelling results and both control and intervention clusters, computes data using single instance of crt_combined and calculates power

power_compute(
  mainpop_data = list(),
  n_clusters = 1,
  n_patients = 1,
  data_level = "Cluster",
  alpha = 0.05,
  effect_size_sign = "Positive",
  trial_type = "sup",
  delta_eq = NULL,
  delta_ni = NULL,
  n_sims = 1,
  benchmark = "EIR_annual",
  age_start = 0,
  age_end = 65,
  test_time_values = c(0, 1),
  test_type = "RDT",
  flag_pre_clearing = 0,
  censor_period = 0,
  flag_reactive_treatment = 1,
  prop_T_c = 0.9,
  benchmark_mean = 0,
  benchmark_stdev = 0,
  int_mean = 0,
  int_stdev = 0
)

Arguments

mainpop_data
List containing main population data output by mainpop()

n_clusters
Number of clusters (control and intervention)

n_patients
Number of patients per cluster

data_level
Data level to use ("Cluster" or "Individual")

alpha
Significance level

effect_size_sign
Sign of effect to check for ("Positive" or "Negative")

trial_type
Trial type (superiority, equivalence, non-inferiority - "sup", "eq", "ni")

delta_eq
Delta for equivalence trials

delta_ni
Delta for non-inferiority trials

n_sims
Number of iterations

benchmark
Benchmark type to use in choosing clusters ("EIR", "slide_prev", "pcr_prev", or "clin_inc" for new data, "EIR_annual", "slide_prev_annual", "pcr_prev_annual", "clin_inc_annual" for pre-existing annual data in input folder)

age_start
Starting age to use when calculating prevalence or incidence over age range (not used with EIR)

age_end
End age to use when calculating prevalence or incidence over age range (not used with EIR)

test_time_values
Time points at which tests are administered

test_type
Type of test administered ("clin" = clinical, "RDT" = rapid diagnostic test)

flag_pre_clearing
Integer indicating whether patients given pre-trial prophylaxis (if 1, place all patients into prophylaxis category at start)

censor_period
Time period after a positive test during which a patient is not counted towards incidence

flag_reactive_treatment
Integer indicating whether patients are automatically given prophylaxis after a positive test (shifting them into treatment category if a clinical case, prophylaxis category otherwise)

prop_T_c
Treatment probability

benchmark_mean
Mean of benchmark value distribution

benchmark_stdev
Standard deviation of benchmark value distribution

int_mean
Mean of intervention parameter value distribution

int_stdev
Standard deviation of intervention parameter value distribution

Details

Takes a list of parameters, returns inf_boolen.